whytebox-v2/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# Cache & Queue
redis==5.0.1
celery==5.3.6

# ML Frameworks
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
timm==0.9.12
transformers==4.36.2

# Model Conversion & Optimization
onnx==1.15.0
onnxruntime==1.16.3
tf2onnx==1.16.1
onnx-tf==1.10.0
tensorflowjs==4.15.0

# Image Processing
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
psutil==5.9.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0